test-tpt: test share pointers, particularly objects shared through pointers of different types
Currently, classes declared with the ClassTP*() macros must always be shared
with TPt pointers that share the same type.  Ie if there is a Child class that
inherits from a Parent class, any individual Child instance must always be
owned/managed by a set of shared pointers of type TPt<Child> *or* TPt<Parent>,
but not a combination of the two.

This is problematic because GLib encourages createing objects exclusively using
a static New() member function, which, by custom, always produces a shared
pointer based off the most-derived type.  This means that pointer cannot be
passed into generic interfaces that want to be able to handle any objects of
an inherited type.